Update: Rename package from lmf_gui to ivfit :tada:
